+++ source/gear/corrosion-harmful.js
@@ -4,18 +4,18 @@ const { joinAsStatement } = require("../util/textUtil");
 
 module.exports = new GearTemplate("Harmful Corrosion",
 	"Inflict @{damage} @{element} damage and @{mod0Stacks} @{mod0} on a foe",
-	"Also inflict @{foeStagger}",
+	"Inflict @{bonus} more Stagger",
 	"Spell",
 	"Fire",
 	350,
 	(targets, user, isCrit, adventure) => {
-		const { element, modifiers: [powerDown], stagger, damage } = module.exports;
+		const { element, modifiers: [powerDown], bonus, damage } = module.exports;
 		let pendingDamage = user.getPower() + damage;
 		if (user.element === element) {
 			changeStagger(targets, "elementMatchFoe");
 		}
 		if (isCrit) {
-			changeStagger(targets, stagger);
+			changeStagger(targets, bonus);
 		}
 		const poweredDownTargets = addModifier(targets, powerDown);
 		if (poweredDownTargets.length > 0) {
@@ -27,6 +27,6 @@ module.exports = new GearTemplate("Harmful Corrosion",
 ).setTargetingTags({ type: "single", team: "foe", needsLivingTargets: true })
 	.setSidegrades("Fate-Sealing Corrosion", "Shattering Corrosion")
 	.setModifiers({ name: "Power Down", stacks: 20 })
-	.setStagger(2)
+	.setBonus(2) // Crit Stagger
 	.setDamage(50)
 	.setDurability(15);