+++ source/gear/spear-reactive.js
@@ -2,33 +2,32 @@ const { GearTemplate, Move } = require('../classes');
 const { dealDamage, changeStagger, getNames } = require('../util/combatantUtil.js');
 
 module.exports = new GearTemplate("Reactive Spear",
-	"Strike a foe for @{damage} (+@{bonus} if after foe) @{element} damage",
-	"Also inflict @{foeStagger}",
+	"Strike a foe for @{damage} (x@{critMultiplier} if after foe) @{element} damage", //TODONOW needs 2 bonuses?
+	"Inflict @{bonus} more Stagger",
 	"Weapon",
 	"Earth",
 	350,
 	([target], user, isCrit, adventure) => {
-		const { element, stagger, damage, bonus } = module.exports;
+		const { element, bonus, damage, critMultiplier } = module.exports;
 		let pendingDamage = user.getPower() + damage;
 		const userMove = adventure.room.moves.find(move => move.userReference.team === user.team && move.userReference.index === adventure.getCombatantIndex(user));
 		const targetMove = adventure.room.moves.find(move => move.userReference.team === target.team && move.userReference.index === adventure.getCombatantIndex(target));
 
 		if (Move.compareMoveSpeed(userMove, targetMove) > 0) {
-			pendingDamage += bonus;
+			pendingDamage *= critMultiplier;
 		}
 		if (user.element === element) {
 			changeStagger([target], "elementMatchFoe");
 		}
 		let resultText = dealDamage([target], user, pendingDamage, false, element, adventure);
 		if (isCrit) {
-			changeStagger([target], stagger);
+			changeStagger([target], bonus);
 			resultText += ` ${getNames([target], adventure)[0]} is Staggered.`;
 		}
 		return resultText;
 	}
 ).setTargetingTags({ type: "single", team: "foe", needsLivingTargets: true })
 	.setSidegrades("Lethal Spear", "Sweeping Spear")
-	.setStagger(2)
 	.setDurability(15)
 	.setDamage(65)
-	.setBonus(75); // damage
+	.setBonus(2); // Crit Stagger