+++ source/gear/warhammer-reactive.js
@@ -2,7 +2,7 @@ const { GearTemplate } = require('../classes');
 const { dealDamage, changeStagger } = require('../util/combatantUtil.js');
 
 module.exports = new GearTemplate("Reactive Warhammer",
-	"Strike a foe for @{damage} (+@{bonus} if foe is stunned) @{element} damage",
+	"Strike a foe for @{damage} (+@{bonus} if foe is stunned, x@{critMultiplier} if after foe) @{element} damage",
 	"Damage x@{critMultiplier}",
 	"Weapon",
 	"Earth",
@@ -17,7 +17,7 @@ module.exports = new GearTemplate("Reactive Warhammer",
 		const targetMove = adventure.room.moves.find(move => move.userReference.team === target.team && move.userReference.index === adventure.getCombatantIndex(target));
 
 		if (compareMoveSpeed(userMove, targetMove) > 0) {
-			pendingDamage += bonus;
+			pendingDamage *= critMultiplier;
 		}
 
 		if (user.element === element) {